generate.py: Fix appending suffix
In the line that was changed, we appended a suffix to `out_dir / name`
where `name` is already a stem, meaning `name` doesn't contain a file
extension. Python's `with_suffix` looks for the last `.` in the given
string and considers everything beyond that an existing suffix. So,
because we were calling `with_suffix` on a stem (without an extension),
`with_suffix` would delete everything past the last `.` in the `name`.
This was incorrect because it didn't correspond to a suffix.