archive: Start archiving posts
Web rot is a common problem: over time, sites that you once
linked to disappear from the web and, thus, that post that
linked to these sites loses meaning. I want to ensure that
this doesn't happen to this blog so I will start to include
static copies of the sites that I link to along with the
actual links. If a link breaks, you can always look at
the archived version of the page.

I won't archive things like links to Wikipedia that aren't
key to the post and contain pretty generic information that
can easily be found elsewhere (also Wikipedia will likely
outlive this blog). Another example of this is programming
language documentation.

I also can't archive links to code on GitHub, but I will
archive the code (later) as plain text documents.

Archives are created with the monolith tool.