Delete personal posts
I don't want to make this blog about personal matters. In fact,
I have come to see the trend of making everything _more personal_
as problematic because it tends to mix the personal and the
professional. The separation of these spheres of life exists
for good reason, giving people the freedom to do whatever they
want in their private lives without affecting their careers.
Essentially, it's a question of personal freedom.